requirements.txt
fastapi
pydantic
pathlib
dataclasses
pyyaml
numpy
pandas
scikit-learn